requirements.txt
streamlit==1.24.1
streamlit-option-menu==0.3.6
pandas==1.4.2
pillow==9.4.0
plotly-express==0.4.1
